Results for Strategy bbrsis
Despite the overall market value decreasing by 44.56%, this strategy made the test portfolio 81.92% in net profit over the course of 8 months. With a drawdown value of 102.39%, this strategy can be optimized in future renditions to potential trade under an 80% drawdown rate, which would make this strategy highly optimal for live trading.